+++ packages/api-server/src/db/Regulation.ts
@@ -1,9 +1,6 @@
 import promiseAll from '@hugsmidjan/qj/promiseAllObject';
 import { execute as htmldiff } from '@island.is/regulations-tools/htmldiff-js';
-import {
-  dePrettify,
-  extractAppendixesAndComments,
-} from '@island.is/regulations-tools/textHelpers';
+import { extractAppendixesAndComments } from '@island.is/regulations-tools/textHelpers';
 import { nameToSlug, toISODate } from '@island.is/regulations-tools/utils';
 import { readFileSync } from 'fs';
 import { FindAttributeOptions, Op, QueryTypes } from 'sequelize';
@@ -74,10 +71,6 @@ async function getRegulationByName(
       attributes,
       where: { name },
     })) ?? undefined;
-
-  if (regulation) {
-    regulation.text = dePrettify(regulation.text);
-  }
   return regulation;
 }
 
@@ -176,10 +169,6 @@ async function getLatestRegulationChange(
         ['id', 'ASC'],
       ],
     })) ?? undefined;
-
-  if (regulationChange) {
-    regulationChange.text = dePrettify(regulationChange.text);
-  }
   return regulationChange;
 }